Created panel for buttons (better sorting) changed mouse listeners to be only at canva panel

+++ net/kyrma/app_drawer/usefull/Vector2d.java
@@ -1,27 +1,27 @@
 package usefull;
 
 public class Vector2d {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
-    public Vector2d(double x, double y) {
+    public Vector2d(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
